make all tests pass on Python 2.5
The default charset being latin9 is not available on Python 2.5
-> it looked like an quite arbitrary choice anyway -> use latin1,
which is a much less arbitrary choice

The with statement is available on Python 2.5 by from __future__
import with_statement only. But instead of doing this we go for
a try-finally block and fix the seek logic in case of failures.
(As a side effect the save_as method should work on even older
Python releases, which however don't seem to be targeted by
multipart anyway, as other language constructs (x if y else z)
are used being available on Python 2.5+ only.)