Fix ResourceWarnings on Python 3
Add a MultipartPart.close method, which consumers can use to close files
(if any) when they're finished with them.  Call it on some error paths
when parts are about to go out of scope, avoiding ResourceWarnings in
the test suite.

Also use a context manager in MultipartPart.save_as to avoid another
more straightforward ResourceWarning.