Use NamedTemporaryFile for large file uploads
I'm considering porting zope.publisher from cgi.FieldStorage to
multipart (see
https://github.com/zopefoundation/zope.publisher/issues/39).  This seems
reasonably practical, except that zope.publisher currently relies on
subclassing cgi.FieldStorage with a make_file method that returns a
NamedTemporaryFile rather than a TemporaryFile.  There doesn't seem a
particular reason why multipart couldn't just use a NamedTemporaryFile
instead.

zope.publisher still supports Python 2, so for Python 2 compatibility
it'll be necessary to monkey-patch this (unless you're willing to
maintain a 0.1.x branch), but in the long term it would be good to be
able to drop the monkey-patch.