Consistently decode URL-encoded form data
It's confusing for raw bytes and percent-encoded bytes in URL-encoded
form data to be decoded using different character sets; this happened if
passing a `charset` parameter other than "latin1" to `parse_form_data`.

https://url.spec.whatwg.org/#application/x-www-form-urlencoded
intentionally doesn't cover non-UTF-8 cases, but it explicitly says that
a parser should perform bytewise percent-decoding followed by UTF-8
decoding, so we might infer that we should do something analogous for
whatever `charset` is specified.  However, Python 2's `unquote` (called
by `parse_qs` unconditionally decodes as ISO-8859-1 if called with
Unicode input that contains percent-encoded sequences.  To work around
this, make sure to call it with bytes input, and then decode the
individual values ourselves.

This is similar to #30, with corresponding test cases, but the details
are different for Python 2.  This was noticed by the Launchpad test
suite after applying the changes from
https://github.com/zopefoundation/zope.publisher/pull/66 (although it
was a problem before that, just masked).