fix: Handle content_type and charset as documented
MultipartSegment.content_type and MS.charset should be None if not defined, as documented.

MultipartPart.content_type should default to "application/octet-stream" for files and "text/plain" for fields, as per RFC 7578. This is not strictly backwards compatible but the correct behavior, so this counts as a fix, not a change.