refactor:  Simpler first boundary check
The index=1 and index>=2 cases can be merged because "buffer[-1,1]" will return an empty slice, so "buffer[index-2, index] != CRLF" correctly detects bad stuff in front of the first boundary for both cases. This saves us a couple of ops on the good path and is also a bit easier to read.